# internal/pkg/http/api/v1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Provides an API for interacting with the server",
        "title": "API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/addRoute": {
            "post": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "add a new route",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Add route",
                "parameters": [
                    {
                        "type": "string",
                        "description": "absolute URL path, e.g. /test or /test.jpg",
                        "name": "urlPath",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "list of b64 encoded HTTP methods, e.g. GET,POST,PUT",
                        "name": "methods",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "list of b64 encoded headers separated by \\r\\n",
                        "name": "headers",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "base64 encoded body",
                        "name": "body",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid Token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/deleteRoute": {
            "post": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "resets a route to default",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Delete route",
                "parameters": [
                    {
                        "type": "string",
                        "description": "absolute URL path, e.g. /test or /test.jpg",
                        "name": "urlPath",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "list of b64 encoded HTTP methods, e.g. GET,POST,PUT",
                        "name": "methods",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid Token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Show the status of server.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "get server metrics in prometheus format",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Get metrics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid Token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/showRoutes": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "description": "show all added routes",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Show routes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid Token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}